+++ R/favored.R
@@ -5,43 +5,44 @@
 #' @param x An object from \code{\link{elo.run}} or \code{\link{elo.glm}}, or for the default method
 #'   a vector representing wins.A.
 #' @param p.A A vector of predicted win probabilities.
-#' @param running logical, denoting whether to use the running predicted values.
-#' @param ... Other arguments (not in use at this time).
+#' @inheritParams mse
 #' @name favored.elo
 NULL
 #> NULL
 
 #' @rdname favored.elo
 #' @export
-favored <- function(x, ...)
+favored <- function(x, ..., subset = TRUE)
 {
   UseMethod("favored")
 }
 
 #' @rdname favored.elo
 #' @export
-favored.elo.run <- function(x, ...)
+favored.elo.run <- function(x, ..., subset = TRUE)
 {
-  favored.default(x$elos[, sum(x$n.players) + 2], fitted(x))
+  favored.default(x$elos[subset, sum(x$n.players) + 2], fitted(x)[subset])
 }
 
 truetiefalse <- function(x) factor(x, levels = c(1, 0.5, 0), labels = c("TRUE", "(tie)", "FALSE"))
 
 #' @rdname favored.elo
 #' @export
-favored.elo.glm <- function(x, ...)
+favored.elo.glm <- function(x, ..., subset = TRUE)
 {
-  if(x$outcome == "score") return(favored.default(x$y, x$fitted.values))
-  table(truetiefalse(score(x$fitted.values, 0)),
-        truetiefalse(score(x$y, 0)), dnn = c("Favored", "Actual"))
+  if(x$outcome == "score") return(favored.default(x$y[subset], x$fitted.values[subset]))
+  table(truetiefalse(score(x$fitted.values[subset], 0)),
+        truetiefalse(score(x$y[subset], 0)), dnn = c("Favored", "Actual"))
 }
 
 #' @rdname favored.elo
 #' @export
-favored.elo.running <- function(x, running = TRUE, ...)
+favored.elo.running <- function(x, running = TRUE, discard.skipped = FALSE, ..., subset = TRUE)
 {
   if(!running) return(NextMethod())
-  favored.default(x$y, x$running.values)
+  if(!is.logical(subset)) stop("'subset' must be logical for this functionality")
+  idx <- attr(object$running.values, "group") > (if(discard.skipped) 0 else -1)
+  favored.default(x$y[idx & subset], x$running.values[idx & subset])
 }
 
 #' @rdname favored.elo