+++ R/mse.R
@@ -2,9 +2,10 @@
 #'
 #' Calculate the mean square error (Brier score) for a model.
 #' @param object An object
-#' @param subset (optional) A vector of indices on which to calculate the MSE.
+#' @param subset (optional) A vector of indices on which to calculate
 #' @param running logical, denoting whether to use the running predicted values.
-#' @param ... Other arguments (not in use at this time).
+#' @param discard.skipped Logical, denoting whether to ignore the skipped observations in the calculation
+#' @param ... Other arguments (not used at this time).
 #' @details Even though logistic regressions don't use the MSE on the y=0/1 scale, it can still be informative.
 #'   Note that the S3 method is \code{mse}.
 #' @name elo.mse
@@ -13,44 +14,40 @@ NULL
 
 #' @rdname elo.mse
 #' @export
-brier <- function(object, subset, ...)
+mse <- function(object, ..., subset = TRUE)
 {
   UseMethod("mse")
 }
 
 #' @rdname elo.mse
 #' @export
-mse <- function(object, subset, ...)
-{
-  UseMethod("mse")
-}
+brier <- mse
 
 #' @rdname elo.mse
 #' @export
-mse.elo.run <- function(object, subset, ...)
+mse.elo.run <- function(object, ..., subset = TRUE)
 {
   r <- residuals(object)
-  if(!missing(subset)) r <- r[subset]
-  mean(r^2)
+  mean(r[subset]^2)
 }
 
 #' @rdname elo.mse
 #' @export
-mse.elo.glm <- function(object, subset, ...)
+mse.elo.glm <- function(object, ..., subset = TRUE)
 {
   r <- object$fitted.values - object$y
-  if(!missing(subset)) r <- r[subset]
-  mean(r^2)
+  mean(r[subset]^2)
 }
 
 #' @rdname elo.mse
 #' @export
-mse.elo.running <- function(object, subset, running = TRUE, ...)
+mse.elo.running <- function(object, running = TRUE, discard.skipped = FALSE, ..., subset = TRUE)
 {
   if(!running) return(NextMethod())
+  if(!is.logical(subset)) stop("'subset' must be logical for this functionality")
+  idx <- attr(object$running.values, "group") > (if(discard.skipped) 0 else -1)
   r <- object$running.values - object$y
-  if(!missing(subset)) r <- r[subset]
-  mean(r^2)
+  mean(r[idx & subset]^2)
 }
 
 #' @rdname elo.mse